Remove .NET 4.5, 4.6, 4.7, 4.7.1 and .NET Core 2.0 targets. Simplify and speedup object instanciation. Restore usage of the string table and mess around with Span<T>

+++ DBClientFiles.NET/Internals/Versions/WDC1.cs
@@ -217,7 +217,7 @@ namespace DBClientFiles.NET.Internals.Versions
             var memberInfo = Members[memberIndex];
 
             var palletOffset = recordReader.ReadBits(memberInfo.OffsetInRecord, memberInfo.BitSize);
-            return _palletTable.Read<T>(palletOffset);
+            return _palletTable.Read<T>((int)palletOffset);
         }
 
         public virtual RecordReader GetRecordReader(int recordSize)
